# openapi.yaml
openapi: 3.0.3
info:
  title: Benchmark Justice™ REST API
  description: Public law analytics platform REST API specification for judges (BJI), prosecutors (PDI), and legislators (LII).
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Local development backend server
  - url: /api
    description: Frontend relative API proxy
paths:
  /search:
    get:
      summary: Search indexed legal and political entities
      description: Returns a list of judges, prosecutors, and legislators matching query criteria.
      parameters:
        - name: q
          in: query
          description: Search text query matching entity names or IDs
          required: false
          schema:
            type: string
        - name: type
          in: query
          description: Filter by entity type
          required: false
          schema:
            type: string
            enum: [judge, prosecutor, legislator]
        - name: state
          in: query
          description: Filter by US state code (e.g., NY, CA, TX) or 'US' for federal
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchItem'
        '503':
          description: Backend service unavailable
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SearchItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [judge, prosecutor, legislator]
        display_name:
          type: string
        state:
          type: string
        current_score:
          type: number
          format: float
...